+++ CHANGELOG.md
@@ -1,9 +1,13 @@
+## next
+
+- Fixed dark mode styles of action buttons in `struct` view for touch devices
+
 ## 1.0.0-beta.47 (02-11-2020)
 
 - Added view inspector
 - Changed nav block to be fixed positioned
 - Changed nav block class name `.discovery-content-badges` -> `.discovery-nav`
-- Added `.overscrolled` class name that's adding to a page element when page's content scrolled over viewport top side
+- Added `.page_overscrolled` class name that's adding to a page element when page's content scrolled over viewport top side
 - Added `page-header` view
 - Added `passive` option support detection test and `passiveCaptureOptions` for `addEventListener()` in `utils.dom`
 - Added `position` option for `popup` view with possible values `trigger` (by default) and `pointer` (to set position according to pointer coordinates)