+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## next
 
 - Fixed dark mode styles of action buttons in `struct` view for touch devices
-- Fixed context for nav to include route attributes
+- Fixed context for nav to include route attributes (#70)
+- Added `nav.remove(name)` method to remove nav items (#72)
 
 ## 1.0.0-beta.47 (02-11-2020)