+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Removed try/catch in `Widget#queryBool()`, so now `when` and `whenData` throws instead of silently fail (#69)
 - Fixed missing horizontal scrolling for `tree` view
 - Fixed postfix style in `badge` view when used in `page-header`'s prelude
+- Fixed `source` view crashing on highlighting
 
 ## 1.0.0-beta.47 (02-11-2020)