+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Fixed postfix style in `badge` view when used in `page-header`'s prelude
 - Fixed `source` view crashing on highlighting
 - Fixed path generation in `signature` view for objects in dictionary mode
+- Allowed to pass a space separated strings to `className`
+- Changed behaviour for a function as `className` value, it's return value can be a string or an array (any other is ignored)
 
 ## 1.0.0-beta.47 (02-11-2020)