+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Allowed to pass a space separated strings to `className`
 - Changed behaviour for a function as `className` value, it's return value can be a string or an array (any other is ignored)
 - Improved output view as a function in view inspector
+- Fixed view's tree in inspector for components with no a container
 
 ## 1.0.0-beta.47 (02-11-2020)