+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## next
+## 1.0.0-beta.48 (04-11-2020)
 
 - Fixed dark mode styles of action buttons in `struct` view for touch devices
 - Fixed context for nav to include route attributes (#70)